[Smoke Tests] Move helper method into utils.
